tools/requirements.txt
# Gemini API (公式推奨SDK)
google-genai>=1.0.0
